Refactor image handling in visualization functions to ensure images are converted to float before processing (fixes the problem with bfloat16 tensors)

+++ src/lightly_train/_visualize/image_classification.py
@@ -38,7 +38,7 @@ def plot_image_classification_labels(
         A single PIL image containing up to max_images annotated images arranged
         in a grid.
     """
-    images = batch["image"].cpu()
+    images = batch["image"].float().cpu()
     gt_classes = [c.cpu() for c in batch["classes"]]
     n = min(max_images, images.shape[0])
