+++ src/lightly_train/_visualize/panoptic_segmentation.py
@@ -52,7 +52,9 @@ def plot_panoptic_segmentation_labels(
     images = batch["image"]
     masks = batch["masks"]
     gt_images = (
-        [img.cpu() for img in images] if isinstance(images, list) else images.cpu()
+        [img.float().cpu() for img in images]
+        if isinstance(images, list)
+        else images.float().cpu()
     )
     gt_masks = [m.cpu() for m in masks] if isinstance(masks, list) else masks.cpu()
     n = min(max_images, len(gt_images))
@@ -122,7 +124,9 @@ def plot_panoptic_segmentation_predictions(
     """
     images = batch["image"]
     gt_images = (
-        [img.cpu() for img in images] if isinstance(images, list) else images.cpu()
+        [img.float().cpu() for img in images]
+        if isinstance(images, list)
+        else images.float().cpu()
     )
     n = min(max_images, len(gt_images), len(pred_masks))
